pyproject.toml
[tool.poetry]
name = "Chat App"
version = "0.1.0"
description = "Chat web app on Websockets"
authors = ["Andrii Sydorenko <sidorenkoandrij217@gmail.com>",]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
isort = "^5.13.2"
flake8 = "^7.1.1"
black = "^24.8.0"
fastapi = {extras = ["standard"], version = "^0.114.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
uvicorn = "^0.30.6"
pre-commit = "^3.8.0"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
colorlog = "^6.8.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
psycopg2-binary = "^2.9.10"

[tool.black]
line-length = 119
target-version = ['py312']
include = '\.py$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.alembic
)/
'''

[tool.isort]
profile = "black"

#[tool.mypy]
#python_version = "3.12"
#exclude = ['alembic/.*']
#plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
#disallow_untyped_calls = true
#ignore_missing_imports = true
